Revert "modified aliasByNode to remove function cruft from metric names "
The patch destroyed the functionality of aliasByNode, which is evident
by the fact that it no longer uses the nodes parameter which is
supposed to influence the outcome of the function.

With this reverted, the function works as described again.

This partially reverts commit 56f9456e4d05a0f2332e48cc51539a76d04caeae.